Removed Nodejs Docker script because it was a PITA
I tried to get Docker to compile Nodejs assets to keep everything in containers
but that proved to be futile. It slowed things down. The build pipeline was
too complex. And there was a bunch of path issues. Maybe this can be done
better in the future, but for now, it provided no benefits.